+++ frontend/app/src/pages/settings/general/index.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import AmountSettingsCategory from '@/components/settings/AmountSettingsCategory.vue';
+import BackendServiceSettingsCategory from '@/components/settings/BackendServiceSettingsCategory.vue';
 import SettingsPage from '@/components/settings/controls/SettingsPage.vue';
 import ExternalServiceSettingsCategory from '@/components/settings/ExternalServiceSettingsCategory.vue';
 import GeneralSettingsCategory from '@/components/settings/GeneralSettingsCategory.vue';
@@ -19,6 +20,7 @@ enum Category {
   AMOUNT = 'amount',
   NFT = 'nft',
   EXTERNAL_SERVICE = 'external-service',
+  BACKEND = 'backend',
 }
 
 const navigation = [
@@ -26,6 +28,7 @@ const navigation = [
   { id: Category.AMOUNT, label: t('general_settings.amount.title') },
   { id: Category.NFT, label: t('general_settings.nft_setting.title') },
   { id: Category.EXTERNAL_SERVICE, label: t('general_settings.external_service_setting.title') },
+  { id: Category.BACKEND, label: t('backend_settings.title') },
 ];
 </script>
 
@@ -38,6 +41,7 @@ const navigation = [
     <AmountSettingsCategory :id="Category.AMOUNT" />
     <NftSettingsCategory :id="Category.NFT" />
     <ExternalServiceSettingsCategory :id="Category.EXTERNAL_SERVICE" />
+    <BackendServiceSettingsCategory :id="Category.BACKEND" />
   </SettingsPage>
 </template>